typeid: extend Sentence and RichSentence
This commit extends both Sentence and RichSentence so they can contain more metadata. This metadata is solely additive so it should not break older drivers or old cache files.

In the future I would suggest renaming Sentence to Token and extract the messages and goals fields into their own tuple, probably named Sentence. But this would be a breaking change.